Proper location of oppimääräSuoritettu
