+++ framework/src/main/java/fi/otavanopisto/pyramus/koski/model/lukio/ops2019/KoskiInternetixLukioStudentHandler2019.java
@@ -66,6 +66,7 @@ public class KoskiInternetixLukioStudentHandler2019 extends AbstractKoskiLukioSt
     boolean laskeKeskiarvot = lopetusSyy != null ? lopetusSyy.getLaskeAinekeskiarvot() : false;
     boolean sisällytäVahvistus = lopetusSyy != null ? lopetusSyy.getSisällytäVahvistaja() : false;
     boolean lukionOppimääräSuoritettu = lopetusSyy != null ? lopetusSyy.isLukionOppimääräSuoritettu() : false;
+    opiskeluoikeus.setOppimaaraSuoritettu(lukionOppimääräSuoritettu || sisällytäVahvistus);
 
     String departmentIdentifier = settings.getToimipisteOID(student.getStudyProgramme().getId(), academyIdentifier);
     
@@ -78,7 +79,6 @@ public class KoskiInternetixLukioStudentHandler2019 extends AbstractKoskiLukioSt
     LukionOppiaineenOppimaaranSuoritus2019 suoritus = new LukionOppiaineenOppimaaranSuoritus2019(
         LukionOppimaara.aikuistenops, Kieli.FI, toimipiste, getDiaarinumero(student));
     suoritus.setTodistuksellaNakyvatLisatiedot(getTodistuksellaNakyvatLisatiedot(student));
-    suoritus.setOppimaaraSuoritettu(lukionOppimääräSuoritettu);
     if (sisällytäVahvistus) {
       suoritus.setVahvistus(getVahvistus(student, departmentIdentifier));
     }
