+++ framework/src/main/java/fi/otavanopisto/pyramus/koski/model/lukio/ops2019/KoskiLukioStudentHandler2019.java
@@ -76,6 +76,7 @@ public class KoskiLukioStudentHandler2019 extends AbstractKoskiLukioStudentHandl
     StudyEndReasonMapping lopetusSyy = opiskelujaksot(student, opiskeluoikeus.getTila(), opintojenRahoitus);
     boolean laskeKeskiarvot = lopetusSyy != null ? lopetusSyy.getLaskeAinekeskiarvot() : false;
     boolean sisällytäVahvistus = lopetusSyy != null ? lopetusSyy.getSisällytäVahvistaja() : false;
+    opiskeluoikeus.setOppimaaraSuoritettu(sisällytäVahvistus);
 
     String departmentIdentifier = settings.getToimipisteOID(student.getStudyProgramme().getId(), academyIdentifier);
