+++ framework/src/main/java/fi/otavanopisto/pyramus/koski/model/lukio/ops2019/LukionOppiaineenOppimaaranSuoritus2019.java
@@ -3,7 +3,6 @@ package fi.otavanopisto.pyramus.koski.model.lukio.ops2019;
 import java.util.HashSet;
 import java.util.Set;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 
@@ -23,15 +22,6 @@ public class LukionOppiaineenOppimaaranSuoritus2019 extends LukionSuoritus2019 {
     this.koulutusmoduuli.setPerusteenDiaarinumero(perusteenDiaarinumero);
   }
   
-  @JsonProperty("oppimääräSuoritettu")
-  public boolean isOppimaaraSuoritettu() {
-    return oppimaaraSuoritettu;
-  }
-
-  public void setOppimaaraSuoritettu(boolean oppimaaraSuoritettu) {
-    this.oppimaaraSuoritettu = oppimaaraSuoritettu;
-  }
-
   public void addOsasuoritus(LukionOppiaineenSuoritus2019 osasuoritus) {
     osasuoritukset.add(osasuoritus);
   }
@@ -45,6 +35,5 @@ public class LukionOppiaineenOppimaaranSuoritus2019 extends LukionSuoritus2019 {
   }
 
   private final LukionOppiaineenOppimaaranKoulutusmoduuli2019 koulutusmoduuli = new LukionOppiaineenOppimaaranKoulutusmoduuli2019();
-  private boolean oppimaaraSuoritettu;
   private final Set<LukionOppiaineenSuoritus2019> osasuoritukset = new HashSet<>();
 }
